feat(playable-adapter-core): support jszip for cocos 2.4.x
