+++ packages/playable-adapter-core/src/channels/facebook/2x.ts
@@ -1,11 +1,11 @@
 import { removeXMLHttpRequest } from "@/utils"
-import { exportDirZipFormSingleFile } from "@/exporter/2x"
+import { exportDirZipFromSingleFile } from "@/exporter/2x"
 import { TChannel, TChannelPkgOptions } from "@/typings"
 
 export const export2xFacebook = async (options: TChannelPkgOptions) => {
   const channel: TChannel = 'Facebook'
 
-  await exportDirZipFormSingleFile({
+  await exportDirZipFromSingleFile({
     ...options,
     channel,
     transformScript: async ($) => {
