+++ packages/playable-adapter-core/src/channels/facebook/3x.ts
@@ -1,10 +1,10 @@
-import { exportDirZipFormSingleFile } from "@/exporter/3x"
+import { exportDirZipFromSingleFile } from "@/exporter/3x"
 import { TChannel, TChannelPkgOptions } from "@/typings"
 
 export const export3xFacebook = async (options: TChannelPkgOptions) => {
   const channel: TChannel = 'Facebook'
 
-  await exportDirZipFormSingleFile({
+  await exportDirZipFromSingleFile({
     ...options,
     channel
   })
