+++ packages/playable-adapter-core/src/channels/mintegral/2x.ts
@@ -1,10 +1,10 @@
-import { exportDirZipFormSingleFile } from "@/exporter/2x"
+import { exportDirZipFromSingleFile } from "@/exporter/2x"
 import { TChannel, TChannelPkgOptions } from "@/typings"
 
 export const export2xMintegral = async (options: TChannelPkgOptions) => {
   const channel: TChannel = 'Mintegral'
 
-  await exportDirZipFormSingleFile({
+  await exportDirZipFromSingleFile({
     ...options,
     channel,
   })
