+++ packages/playable-adapter-core/src/channels/mintegral/3x.ts
@@ -1,10 +1,10 @@
-import { exportDirZipFormSingleFile } from "@/exporter/3x"
+import { exportDirZipFromSingleFile } from "@/exporter/3x"
 import { TChannel, TChannelPkgOptions } from "@/typings"
 
 export const export3xMintegral = async (options: TChannelPkgOptions) => {
   const channel: TChannel = 'Mintegral'
 
-  await exportDirZipFormSingleFile({
+  await exportDirZipFromSingleFile({
     ...options,
     channel,
   })
