+++ packages/playable-adapter-core/src/packager/2x.ts
@@ -2,7 +2,7 @@ import {
   TChannel,
   TChannelPkgOptions,
 } from '@/typings'
-import { getAdapterRCJson } from '@/utils'
+import { genChannelsPkg as baseGenChannelsPkg } from './base'
 import {
   export2xAppLovin,
   export2xFacebook,
@@ -32,23 +32,5 @@ const channelExports: { [key in TChannel]: (options: TChannelPkgOptions) => Prom
 }
 
 export const genChannelsPkg = (options: TChannelPkgOptions): Promise<void> => {
-  return new Promise((resolve, reject) => {
-    const { exportChannels = [] } = getAdapterRCJson() || {}
-
-    let tasks: Promise<void>[] = []
-    let channelKeys = exportChannels.length === 0 ? <TChannel[]>Object.keys(channelExports) : exportChannels
-    channelKeys.forEach((key: TChannel) => {
-      if (exportChannels.length === 0 || exportChannels.includes(key)) {
-        tasks.push(channelExports[key](options))
-      }
-    })
-    Promise.all(tasks)
-      .then(() => {
-        resolve()
-      })
-      .catch((err) => {
-        console.error(err)
-        reject(err)
-      })
-  })
+  return baseGenChannelsPkg(channelExports, options)
 }
